move scatter out
